// appekko-ui/src/app/core/enums/languages.ts
export enum ETechnologyLanguage {
    ASSEMBLY = 'Assembly',
    C = 'C',
    COFFEESCRIPT = 'CoffeeScript',
    CPP = 'C++',
    CSHARP = 'C#',
    CSS = 'CSS',
    DART = 'Dart',
    DM = 'DM',
    ELIXIR = 'Elixir',
    GO = 'Go',
    GROOVY = 'Groovy',
    HTML = 'HTML',
    JAVA = 'Java',
    JAVASCRIPT = 'JavaScript',
    KOTLIN = 'Kotlin',
    OBJECTIVEC = 'Objective-C',
    PERL = 'Perl',
    PHP = 'PHP',
    POWERSHELL = 'Powershell',
    PYTHON = 'Python',
    RUBY = 'Ruby',
    RUST = 'Rust',
    SCALA = 'Scala',
    SHELL = 'Shell',
    SWIFT = 'Swift',
    TYPESCRIPT = 'TypeScript'
}

export const TechnologyLanguages = [
    ETechnologyLanguage.ASSEMBLY,
    ETechnologyLanguage.C,
    ETechnologyLanguage.COFFEESCRIPT,
    ETechnologyLanguage.CPP,
    ETechnologyLanguage.CSHARP,
    ETechnologyLanguage.CSS,
    ETechnologyLanguage.DART,
    ETechnologyLanguage.DM,
    ETechnologyLanguage.ELIXIR,
    ETechnologyLanguage.GO,
    ETechnologyLanguage.GROOVY,
    ETechnologyLanguage.HTML,
    ETechnologyLanguage.JAVA,
    ETechnologyLanguage.JAVASCRIPT,
    ETechnologyLanguage.KOTLIN,
    ETechnologyLanguage.OBJECTIVEC,
    ETechnologyLanguage.PERL,
    ETechnologyLanguage.PHP,
    ETechnologyLanguage.POWERSHELL,
    ETechnologyLanguage.PYTHON,
    ETechnologyLanguage.RUBY,
    ETechnologyLanguage.RUST,
    ETechnologyLanguage.SCALA,
    ETechnologyLanguage.SHELL,
    ETechnologyLanguage.SWIFT,
    ETechnologyLanguage.TYPESCRIPT,
]